test: move pure integration tests to unit
