+++ src/Core/Framework/Adapter/Cache/StampedeProtectionConfigurator.php
@@ -19,8 +19,6 @@ use Symfony\Component\Cache\LockRegistry;
  * See https://github.com/shopware/shopware/issues/12823#issuecomment-3677936635
  *
  * @internal
- *
- * @codeCoverageIgnore @see \Shopware\Tests\Integration\Core\Framework\Adapter\Cache\StampedeProtectionConfiguratorTest
  */
 #[Package('framework')]
 readonly class StampedeProtectionConfigurator
