+++ src/Core/Framework/DependencyInjection/CompilerPass/TwigEnvironmentCompilerPass.php
@@ -8,9 +8,6 @@ use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 #[Package('framework')]
-/**
- * @codeCoverageIgnore This would be useless as a unit test. It is integration tested here: \Shopware\Tests\Integration\Core\Framework\DependencyInjection\CompilerPass\TwigEnvironmentCompilerPassTest
- */
 class TwigEnvironmentCompilerPass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container): void
