+++ src/Core/Framework/Util/Filesystem.php
@@ -3,7 +3,6 @@
 namespace Shopware\Core\Framework\Util;
 
 use Shopware\Core\Framework\Log\Package;
-use Shopware\Tests\Integration\Core\Framework\Util\FilesystemTest;
 use Symfony\Component\Filesystem\Filesystem as Io;
 use Symfony\Component\Filesystem\Path;
 use Symfony\Component\Finder\Finder;
@@ -11,10 +10,6 @@ use Symfony\Component\Finder\SplFileInfo;
 
 /**
  * @internal
- *
- * @codeCoverageIgnore
- *
- * @see FilesystemTest
  */
 #[Package('framework')]
 class Filesystem
