+++ tests/unit/Core/Content/Media/Aggregate/MediaThumbnail/MediaThumbnailStructTest.php
@@ -1,13 +1,15 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Content\Media\Aggregate\MediaThumbnail;
+namespace Shopware\Tests\Unit\Core\Content\Media\Aggregate\MediaThumbnail;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Content\Media\Aggregate\MediaThumbnail\MediaThumbnailEntity;
 
 /**
  * @internal
  */
+#[CoversClass(MediaThumbnailEntity::class)]
 class MediaThumbnailStructTest extends TestCase
 {
     public function testGetIdentifier(): void
