+++ tests/unit/Core/Framework/Adapter/Cache/StampedeProtectionConfiguratorTest.php
@@ -1,7 +1,8 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\Adapter\Cache;
+namespace Shopware\Tests\Unit\Core\Framework\Adapter\Cache;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Adapter\Cache\StampedeProtectionConfigurator;
@@ -10,6 +11,7 @@ use Symfony\Component\Cache\LockRegistry;
 /**
  * @internal
  */
+#[CoversClass(StampedeProtectionConfigurator::class)]
 class StampedeProtectionConfiguratorTest extends TestCase
 {
     private const TEST_LOCK_FILES = ['/tmp/test-lock-1'];
