+++ tests/unit/Core/Framework/App/ActionButton/Response/ActionButtonResponseFactoryTest.php
@@ -1,36 +1,50 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App\ActionButton\Response;
+namespace Shopware\Tests\Unit\Core\Framework\App\ActionButton\Response;
 
+use GuzzleHttp\Psr7\Uri;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\ActionButton\AppAction;
 use Shopware\Core\Framework\App\ActionButton\Response\ActionButtonResponseFactory;
 use Shopware\Core\Framework\App\ActionButton\Response\NotificationResponse;
+use Shopware\Core\Framework\App\ActionButton\Response\NotificationResponseFactory;
 use Shopware\Core\Framework\App\ActionButton\Response\OpenModalResponse;
+use Shopware\Core\Framework\App\ActionButton\Response\OpenModalResponseFactory;
 use Shopware\Core\Framework\App\ActionButton\Response\OpenNewTabResponse;
+use Shopware\Core\Framework\App\ActionButton\Response\OpenNewTabResponseFactory;
 use Shopware\Core\Framework\App\ActionButton\Response\ReloadDataResponse;
+use Shopware\Core\Framework\App\ActionButton\Response\ReloadDataResponseFactory;
 use Shopware\Core\Framework\App\AppEntity;
 use Shopware\Core\Framework\App\AppException;
+use Shopware\Core\Framework\App\Hmac\QuerySigner;
 use Shopware\Core\Framework\App\Payload\Source;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Uuid\Uuid;
-use Shopware\Tests\Integration\Core\Framework\App\GuzzleTestClientBehaviour;
 
 /**
  * @internal
  */
+#[CoversClass(ActionButtonResponseFactory::class)]
 class ActionButtonResponseFactoryTest extends TestCase
 {
-    use GuzzleTestClientBehaviour;
-
     private ActionButtonResponseFactory $actionButtonResponseFactory;
 
     private AppAction $action;
 
     protected function setUp(): void
     {
-        $this->actionButtonResponseFactory = static::getContainer()->get(ActionButtonResponseFactory::class);
+        $signer = $this->createMock(QuerySigner::class);
+        $signer->method('signUri')->willReturn(new Uri('http://signed.url'));
+
+        $this->actionButtonResponseFactory = new ActionButtonResponseFactory([
+            new NotificationResponseFactory(),
+            new ReloadDataResponseFactory(),
+            new OpenNewTabResponseFactory($signer),
+            new OpenModalResponseFactory($signer),
+        ]);
+
         $app = new AppEntity();
         $app->setName('TestApp');
         $app->setId(Uuid::randomHex());
