+++ tests/unit/Core/Framework/App/ActionButton/Response/OpenModalResponseFactoryTest.php
@@ -1,36 +1,42 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App\ActionButton\Response;
+namespace Shopware\Tests\Unit\Core\Framework\App\ActionButton\Response;
 
+use GuzzleHttp\Psr7\Uri;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\ActionButton\AppAction;
 use Shopware\Core\Framework\App\ActionButton\Response\NotificationResponse;
 use Shopware\Core\Framework\App\ActionButton\Response\OpenModalResponse;
+use Shopware\Core\Framework\App\ActionButton\Response\OpenModalResponseFactory;
 use Shopware\Core\Framework\App\ActionButton\Response\OpenNewTabResponse;
-use Shopware\Core\Framework\App\ActionButton\Response\OpenNewTabResponseFactory;
 use Shopware\Core\Framework\App\ActionButton\Response\ReloadDataResponse;
 use Shopware\Core\Framework\App\AppEntity;
 use Shopware\Core\Framework\App\AppException;
+use Shopware\Core\Framework\App\Hmac\QuerySigner;
 use Shopware\Core\Framework\App\Payload\Source;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Uuid\Uuid;
-use Shopware\Tests\Integration\Core\Framework\App\GuzzleTestClientBehaviour;
 
 /**
  * @internal
  */
-class OpenNewTabResponseFactoryTest extends TestCase
+#[CoversClass(OpenModalResponseFactory::class)]
+class OpenModalResponseFactoryTest extends TestCase
 {
-    use GuzzleTestClientBehaviour;
-
-    private OpenNewTabResponseFactory $factory;
+    private OpenModalResponseFactory $factory;
 
     private AppAction $action;
 
+    private QuerySigner&MockObject $signer;
+
     protected function setUp(): void
     {
-        $this->factory = static::getContainer()->get(OpenNewTabResponseFactory::class);
+        $this->signer = $this->createMock(QuerySigner::class);
+        $this->factory = new OpenModalResponseFactory($this->signer);
+
         $app = new AppEntity();
         $app->setName('TestApp');
         $app->setId(Uuid::randomHex());
@@ -49,22 +55,30 @@ class OpenNewTabResponseFactoryTest extends TestCase
     }
 
     #[DataProvider('provideActionTypes')]
-    public function testSupportsOnlyOpenNewTabActionType(string $actionType, bool $isSupported): void
+    public function testSupportsOnlyOpenModalActionType(string $actionType, bool $isSupported): void
     {
         static::assertSame($isSupported, $this->factory->supports($actionType));
     }
 
-    public function testCreatesOpenNewTabResponse(): void
+    public function testCreatesOpenModalResponse(): void
     {
-        $response = $this->factory->create($this->action, [
-            'redirectUrl' => 'http://redirect.url',
-        ], Context::createDefaultContext());
+        $context = Context::createDefaultContext();
+        $this->signer->expects($this->once())
+            ->method('signUri')
+            ->with('http://iframe.url', $this->action->getApp(), $context)
+            ->willReturn(new Uri('http://iframe.url?shopware-shop-signature=signature'));
 
-        static::assertInstanceOf(OpenNewTabResponse::class, $response);
+        $response = $this->factory->create($this->action, [
+            'iframeUrl' => 'http://iframe.url',
+            'size' => 'medium',
+            'expand' => false,
+        ], $context);
+
+        static::assertInstanceOf(OpenModalResponse::class, $response);
     }
 
     /**
-     * @param array<string, mixed> $payload
+     * @param array<bool|string> $payload
      */
     #[DataProvider('provideInvalidPayloads')]
     public function testThrowsExceptionWhenValidationFails(array $payload, string $message): void
@@ -86,25 +100,37 @@ class OpenNewTabResponseFactoryTest extends TestCase
     {
         return [
             [NotificationResponse::ACTION_TYPE, false],
-            [OpenModalResponse::ACTION_TYPE, false],
-            [OpenNewTabResponse::ACTION_TYPE, true],
+            [OpenModalResponse::ACTION_TYPE, true],
+            [OpenNewTabResponse::ACTION_TYPE, false],
             [ReloadDataResponse::ACTION_TYPE, false],
         ];
     }
 
     /**
-     * @return array<array<string|array<string, mixed>>>
+     * @return array<array<array<bool|string>|string>>
      */
     public static function provideInvalidPayloads(): array
     {
         return [
             [
-                [],
-                'The app provided an invalid redirectUrl',
+                ['size' => 'medium', 'expand' => false],
+                'The app provided an invalid iframeUrl',
             ],
             [
-                ['redirectUrl' => ''],
-                'The app provided an invalid redirectUrl',
+                ['iframeUrl' => '', 'size' => 'medium', 'expand' => false],
+                'The app provided an invalid iframeUrl',
+            ],
+            [
+                ['iframeUrl' => 'http://iframe.url', 'expand' => false],
+                'The app provided an invalid size',
+            ],
+            [
+                ['iframeUrl' => 'http://iframe.url', 'size' => '', 'expand' => false],
+                'The app provided an invalid size',
+            ],
+            [
+                ['iframeUrl' => 'http://iframe.url', 'size' => 'xl', 'expand' => false],
+                'The app provided an invalid size',
             ],
         ];
     }
