+++ tests/unit/Core/Framework/App/AppArchiveValidatorTest.php
@@ -1,7 +1,8 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App;
+namespace Shopware\Tests\Unit\Core\Framework\App;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\AppArchiveValidator;
 use Shopware\Core\Framework\App\Exception\AppArchiveValidationFailure;
@@ -11,6 +12,7 @@ use Symfony\Component\Filesystem\Filesystem;
 /**
  * @internal
  */
+#[CoversClass(AppArchiveValidator::class)]
 class AppArchiveValidatorTest extends TestCase
 {
     private string $temporaryFilePath;
