+++ tests/unit/Core/Framework/App/AppExtractorTest.php
@@ -1,7 +1,8 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App;
+namespace Shopware\Tests\Unit\Core\Framework\App;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\AppArchiveValidator;
 use Shopware\Core\Framework\App\AppExtractor;
@@ -13,6 +14,7 @@ use Symfony\Component\Filesystem\Path;
 /**
  * @internal
  */
+#[CoversClass(AppExtractor::class)]
 class AppExtractorTest extends TestCase
 {
     private string $temporaryFilePath;
