+++ tests/unit/Core/Framework/App/_fixtures/App/TestApp/icon.png
