+++ tests/unit/Core/Framework/App/_fixtures/AppInvalidPrefix/TestApp/icon.png
