+++ tests/unit/Core/Framework/App/_fixtures/AppInvalidPrefix/TestApp/manifest.xml
