+++ tests/unit/Core/Framework/App/_fixtures/AppWithoutManifestFile/TestApp/icon.png
