+++ tests/unit/Core/Framework/App/_fixtures/FileTraversal.zip
